+++ shared/src/ecs/systems/index.ts
@@ -1,5 +1,5 @@
 export * from './WeaponSystem';
-export * from './VehicleSystems'; 
+export * from './DroneSystems'; 
 export * from './EntitySystem';
 export * from './PhysicsWorldSystem';
 export * from './HealthSystems';
